requirements.txt
torch
transformers>=4.40.0
accelerate>=0.27.0
safetensors
numpy
tqdm
pillow
pyyaml
sentencepiece
